+++ components/tabs/TabBar.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import Icon from '../icon';
 import ScrollableInkTabBar from 'rc-tabs/lib/ScrollableInkTabBar';
 import { TabsProps } from './index';
 
@@ -9,16 +10,35 @@ export default class TabBar extends React.Component<TabsProps> {
       animated = true,
       renderTabBar,
       tabBarExtraContent,
+      tabPosition,
+      prefixCls,
     } = this.props;
     const inkBarAnimated = typeof animated === 'object' ? animated.inkBar : animated;
 
-    let RenderTabBar: React.ReactElement<any>;
+    const isVertical = tabPosition === 'left' || tabPosition === 'right';
+    const prevIconType = isVertical ? 'up' : 'left';
+    const nextIconType = isVertical ? 'down' : 'right';
+    const prevIcon = (
+      <span className={`${prefixCls}-tab-prev-icon`}>
+        <Icon type={prevIconType} className={`${prefixCls}-tab-prev-icon-target`} />
+      </span>
+    );
+    const nextIcon = (
+      <span className={`${prefixCls}-tab-next-icon`}>
+        <Icon type={nextIconType} className={`${prefixCls}-tab-next-icon-target`} />
+      </span>
+    );
+
     const renderProps = {
       ...this.props,
       inkBarAnimated,
       extraContent: tabBarExtraContent,
       style: tabBarStyle,
+      prevIcon,
+      nextIcon,
     };
+
+    let RenderTabBar: React.ReactElement<any>;
 
     if (renderTabBar) {
       RenderTabBar = renderTabBar(renderProps);
