+++ components/tabs/index.tsx
@@ -161,35 +161,6 @@ export default class Tabs extends React.Component<TabsProps, any> {
       </div>
     ) : null;
 
-    const renderTabBar = () => {
-      const isVertical = tabPosition === 'left' || tabPosition === 'right';
-      const prevIconType = isVertical ? 'up' : 'left';
-      const nextIconType = isVertical ? 'down' : 'right';
-      const prevIcon = (
-        <span className={`${prefixCls}-tab-prev-icon`}>
-          <Icon type={prevIconType} className={`${prefixCls}-tab-prev-icon-target`} />
-        </span>
-      );
-      const nextIcon = (
-        <span className={`${prefixCls}-tab-next-icon`}>
-          <Icon type={nextIconType} className={`${prefixCls}-tab-next-icon-target`} />
-        </span>
-      );
-      return (
-        <ScrollableInkTabBar
-          inkBarAnimated={inkBarAnimated}
-          extraContent={tabBarExtraContent}
-          onTabClick={onTabClick}
-          onPrevClick={onPrevClick}
-          onNextClick={onNextClick}
-          style={tabBarStyle}
-          tabBarGutter={tabBarGutter}
-          prevIcon={prevIcon}
-          nextIcon={nextIcon}
-        />
-      );
-    };
-
     return (
       <RcTabs
         {...this.props}
